Add ToolRepository and initial test
