+++ packages/cli/jest.config.agents.js
@@ -1,7 +1,7 @@
 /** @type {import('jest').Config} */
-// Dedicated config for agent integration tests that require the real filesystem.
+// Dedicated config for agent integration tests that load real nodes from disk.
 // setup-mocks.ts is intentionally excluded — it mocks node:fs which breaks
-// LazyPackageDirectoryLoader and any test that loads real node types from disk.
+// LazyPackageDirectoryLoader.
 module.exports = {
 	...require('../../jest.config'),
 	testEnvironmentOptions: {
@@ -11,7 +11,7 @@ module.exports = {
 	globalTeardown: '<rootDir>/test/teardown.ts',
 	setupFilesAfterEnv: ['<rootDir>/test/setup-test-folder.ts', '<rootDir>/test/extend-expect.ts'],
 	testRegex: undefined,
-	testMatch: ['<rootDir>/src/modules/agents/__tests__/*.integration.test.ts'],
+	testMatch: ['<rootDir>/src/modules/agents/__tests__/*.test.ts'],
 	testPathIgnorePatterns: ['/dist/', '/node_modules/'],
 	testTimeout: 30_000,
 	prettierPath: null,
